+++ ft-sys-shared/src/lib.rs
@@ -3,7 +3,10 @@
 //! `ft-sdk` should be used.
 #![deny(unused_extern_crates)]
 
+mod email;
 mod sqlite;
+
+pub use email::{Email, EmailAddress, EmailHandle};
 pub use sqlite::{SqliteRawValue, SqliteType};
 
 pub const SESSION_KEY: &str = "fastn-sid";